Pad the transfer buffer to prevent out of bounds access.
The HIDReportItem reads 32 bit chunks from the report buffer. To avoid
having to check the remaining buffer space on each extraction, we pad
the buffer so it is always valid to read 32 bits at a time. Also add
a comment explaining why we do it that way.

Thanks to Johannes Anderwald for pointing out the potential out of
bound access!