intel accelerant: NULL dereferences (overlays)
Rather than checking if "window" AND "view" are NULL, check if
"window" OR "view" are NULL, since neither could be null and deferenced later
on.

From the Matrox accelerant, it seems that such event occurs when a
"Workspace switch, screen prefs change, or overlay app shutdown occurs"

Might be interesting to revisit tickets related to intel accelerant.

CID 205.