Fix build for !x86 && !ppc
At least this should fix other platform builds to the point they worked before.
The ICU_DEVEL zip name assign was moved inside the if block, making it empty for other archs, and ExtractArchive was of course having issues with this.
So I moved the rest inside the if block as well. Should fix at least kernel builds.