Updated WebPositive. Supports common/data/ssl/cert.pem
This package contains only the app + libs. The certificate file
is satisfied through OptionalPackageDependencies. The bookmarks
have been moved (and updated) to a separate archive.

Thanks Rene Gollent and Chris Roberts for the help!

This group of changesets should close #7707, though it is possible
that other optional-packages, pre-built software will still fail to
locate the certificates file in the new location.