Addresses some of the issues brought up in ticket #8148
The key labels have been changed to Win/Option and Alt/Command
to better associate them with the key names on a PC keyboard. This is
not perfect, but, better.

Added a checkbox entitled "Switch right Alt/Command and Win/Option keys"
This checkbox does what it says, switches the keys on the right side
of your keyboard leaving the left side according to how you set them.
This is needed for some international keymaps.

Whether or not you have switched already is not remembered, you have to
check the box each time you open the dialog. I don't know how to figure
out whether or not the keys have been switched already reliably. Hopefully
this isn't a big deal.

Set the menu option and window title to sentence case, thanks Humdinger.