Add 12/24 hour and time formatting options to Deskbar.
* 12/24 hour settings come from and alter BLocale
* Time formats are short, medium, long. Short is default. Come from
  Locale kit.
* Don't delete fClock (now fTime) when hiding, just hide and show
  the view.
* BarView now has nothing to do with setting or keeping track of
  time settings. This is all handled in TTimeView. TReplicantTray
  is responsible for updating the replicant tray and doing showing
  and hiding. TTimeView is responsible for drawing according to the
  clock settings.
* Remove fShowInterval and fShowSeconds from BarView
* Remove ampmMode setting and replace it with timeFormat.
* Reset targets in DeskbarMenu.

Originally I was trying to solve a bug where the TTimeView
was relying on the Deskbar settings too much. The settings
should only by set on quit and accessed on startup or a crash can
cause Deskbar to be in an unusual state.

I realize that the timezone is not very helpful. I'd like to provide
the day of week in a short format instead, i.e. Wed for Wednesday or
similar based on your locale. Blame the Locale Kit for now allowing
me to be able to use a custom TimeFormat. Once a custom TimeFormats
becomes possible from the Locale Kit Deskbar can be updated to use
them.