Some Layout fixed to 3dRendering for different font sizes.
Add glue between the gears view and the rest to make sure that
the rest butts up against GearsView. Update the Preferred
and Max size so that the spacing is appropriate. Make the window
width smaller and allow the views to expand it.

The ColumnListViews are still not set to respect different font
sizes but overall it works better at smaller font sizes.