Put the InfoView in the main OpenGLView.
Take the InfoView out of the tabView and put it above the tabView
in the OpenGLView. Change the format to a slightly altered version
from what kallisti5 recommended. The width and height of the tabView
are controlled by M's so if you change your font size it will fit
accordingly. Therefore the window dimensions are sane defaults yet
don't really affect the dimensions of the child views. The views
know how to size themselves sanely.

* Change the 3d Rendering Engine from a BMenu to a BPopUpMenu.