Nip this M thing in the bud.
Multiply StringWidth("M") * the number of M's needed and then
tweak to get it to fit well at different font sizes. Turns out
it fits the same most of the time. Tested at font size 12pt, 8pt,
and 18pt.There are issues with BStringColumn at 18pt not
adjusting to the font size correctly but this is a different bug.

Add myself to the authors list on InfoView.cpp since I have
updated the file heavily at this point.