After setting the tooltip text, invalidate the layout.
This fixes a bug introduced in hrev44075 (I think) where the internal
tooltip window wouldn't resize itself after setting new tooltip text.
So if you set the tooltip to a short string then set it to a long
string the tooltip would get cut-off or if you set it to a long string
and then set it to a short string the tooltip window would be too wide.
The result can be seen in Deskbar in the following screenshot:

http://30.media.tumblr.com/tumblr_m3fg327NsV1r0f0hfo1_400.png

After invalidating the layout the internal tooltip window gets resized
correctly.