Fix formatting bug in Deskcalc.
Calculations that take up the full width of the Deskcalc window
sometimes scroll the result over showing you only the tail end. This
is because the horizontal inset makes the area too small to fit the
result. Removing this horizontal inset means that the result will
always fit in the space provided, and Deskcalc will recalculate to
fit, but it also means that the result always starts at the leftmost
side of the textarea, a fair tradeoff.