Fix off-by-one error which caused scrolling artifacts.
Based on printf debugging the correct left value is actually -1, not 0, but
instead of hard-coding numbers it makes more sense to use the existing value.
I'm not sure if a -1 left value makes much sense, but overall this seems to
work and does not have the scrolling artifacts.

But overall I think ShowImage needs more layout management. Doing all this
moving and resizing manually is error prone.

Fixes #8648.