FileTypes : IconView for non-applications fixes
IconView is now only setting the icon with BAppFileInfo when the file is an application,
otherwise relies on BNodeInfo. This avoids adding ressources within data files.

The IconView also wasn't populated correctly when used on data files (by opposition to
application binaries).  It now loads the same icon that is shown in Tracker.