ls: Change color scheme for other writable directory (#4494)
The ls command, uses blue color for directories. For a directory that
has write permission for others, it uses green highlighting while
displaying the directory name. However, blue text and green highlight
is difficult to read.

The patch keeps the text color as blue and changes the highlight to
yellow.

Signed-off-by: Jérôme Duval <jerome.duval@gmail.com>