Preserve the clipping constraints of the parent view...
when drawing the following controls:
* SliderBar
* ActiveTab
* MenuField

This is a followup commit based on the change made for buttons in
hrev44708 in response to bug #8700. It is good practice to always preserve
the parent view's clipping constraints.

Stippi and Axel can you you at this commit and make sure this is kosher?

+alpha4