Draw rounded inner corners for return key in Keymap.
Draw the return key with rounded inner corners instead of square, this
builds on top of the changes made in hrev44709. The secret to rounding
the corners was to draw both the edge and background of the individual
sections before drawing the button's background on top minus the
clipped out bottom left section.

+alpha4 (optionally, purely cosmetic but shouldn't hurt)