Disable duplicated interrupt disable code in marvell yukon.
Since we now implement HAIKU_CHECK_DISABLE_INTERRUPTS there is no need
to do it again in the interrupt handler. Further, store and carry over
the interrupt status from the hook to the interrupt handler as that is
cleaner and just in case reading the register somehow clears it. May
help with #8454.