Workaround possible spurious interrupts in marvell yukon.
This comes from a similar patch in FreeBSD r234666 but does the check
and acknowledge in the case where we don't have any other interrupts.
Since the interrupt isn't used it is never unmasked and therefore the
check against the interrupt mask will always fail if the TWSI IRQ is
the only interrupt that happened. In that case an interrupt storm would
be triggered that lasts until any other valid interrupt comes along, is
handled and both are acknowledged.

This may help with #8454 on some cards that exhibit the spurious TWSI
IRQs.