urlwrapper: Handle the doi: URI scheme
The DOI (Digital Object Identifier) is a unique identifier
for (scientific and other...) publications.
While the doi: URI scheme was only a draft, it's not totally unlikely
that we ever hit such a URI, in this case we just redirect to the
official DOI website.
cf. http://tools.ietf.org/html/draft-paskin-doi-uri