Update Deskbar Clock to use predefined time formats
... with day of week tacked on at the beginning.

This fixes #9143 by better allowing the Locale Kit to
format the time. It was localized before but now also
uses localized time separators.

There might be still a bug with day of week though,
depending on if day of week should go before or after
the time in your locale (It is hard coded to before).