Remove obsolete time update checking code.
This was put into OpenDeskbar a long long time ago (2001):

http://opentracker.cvs.sourceforge.net/viewvc/opentracker/opentracker/deskbar/TimeView.cpp?r1=1.2&r2=1.3

However, this trick will no longer work because we use a localized
time string from the Locale Kit that might not have a ':' as a
time separator anymore. Also with day of week we don't know the
position of the time separator either.